Removed @EnableZeebeClient and switched to AutoConfiguration (also removed lifecycle/proxy for ZeebeClient from production setup, only use lifecycle for annotationRegistry, introduced SpringZeebeTestContext marker bean to exclude it in test cases)
(cherry picked from commit 5de1b150df4f4963c4b2c0ff898b1b0ef7bc7453)
